backend/requirements.txt
alembic==1.4.2
arrow==0.15.6
asn1crypto==1.3.0
bert-serving-client==1.10.0
blinker==1.4
boto==2.49.0
boto3==1.13.19
botocore==1.16.15
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
docutils==0.16
dominate==2.5.1
elasticsearch==7.7.1
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Humanize==0.3.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
gensim==3.8.3
google-api-core==1.17.0
google-auth==1.16.0
google-cloud-core==1.3.0
google-cloud-storage==1.28.1
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
humanize==2.4.0
idna==2.9
IPy==1.0
itsdangerous==1.1.0
jieba-fast==0.53
Jinja2==2.11.2
jmespath==0.10.0
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
numpy==1.18.4
protobuf==3.12.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyMySQL==0.9.3
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
pyzmq==19.0.1
requests==2.23.0
rsa==4.0
s3transfer==0.3.3
scipy==1.4.1
sentry-sdk==0.14.4
six==1.15.0
smart-open==2.0.0
SQLAlchemy==1.3.17
urllib3==1.25.9
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.1
WTForms-JSON==0.3.3
